move the index check exceptions up
